+++ packages/windie-sdk-js/cjs/runtime/WindieAgent.js
@@ -193,7 +193,7 @@ class WindieAgent {
             conversationRef,
             revisionId: options.revisionId,
             store: options.store ?? this.defaultConversationStore,
-            transport: (0, WindieAgentSession_js_1.createWindieAgentBackendTransport)(this.session, conversationRef, this.agentDefinition),
+            transport: (0, WindieAgentSession_js_1.createAgentBackendTransport)(this.session, conversationRef, this.agentDefinition),
             localRuntime: resolvedLocalRuntime,
             sdkClient: this.sdkClient,
             userId: this.userId,